Our submission for the Hackathon.
We gave it our best--and we definitely learned a lot. This was our first
time working with databases and web development. We successfully manage
the data in the database with Python, and we can produce the webpages in
HTML and JavaScript. The webpages also show the code to embed the graph
in another website. Unfortunately, we were not able to successfully link
these two pieces of the project.

We had a blast, and we look forward to TribeHacks next year!